# perf(ops): directly respond for eager ops